+++ include/utils/wtree.h
@@ -1,6 +1,8 @@
 #ifndef WTREE_H
 #define WTREE_H
 
+#include <stddef.h>
+
 #define WTREE_CHARS_NUMBER 64
 #define WTREE_CHARS                                                            \
   "-0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZ_abcdefghijklmnopqrstuvwxyz"
@@ -8,7 +10,7 @@
 struct WTree *create_wtree();
 void expand_wtree(struct WTree *wtree, char *word, void *endpoint);
 void shrink_wtree(struct WTree *wtree, char *word);
-unsigned int get_wtree_size(struct WTree *wtree);
+size_t get_wtree_size(struct WTree *wtree);
 void sort_wtree(struct WTree *wtree);
 void *get_wtree_endpoint(struct WTree *wtree, char *word);
 char **get_wtree_words(struct WTree *wtree);
